refactor: request, response type 분리
